COTECH-1101 rename tagless-request-url-builder.ts after removing buildTaglessRequestUrl()
