+++ src/core/utils/index.ts
@@ -22,10 +22,10 @@ export * from './sampler';
 export * from './script-loader';
 export * from './service-base-url';
 export * from './string-builder';
-export * from './tagless-request-url-builder';
 export * from './targeting';
 export * from './timed-partner-script-loader';
 export * from './try-property';
+export * from './url-builder';
 export * from './uuid';
 export * from './viewport-observer';
 export * from './wait-for';
